interpreter: evaluate symbol data wrapper by resolving in runtime environment
Previously, AST nodes of type AST_TYPE_DATA_WRAPPER with data type TYPE_SYMBOL
were always converted into a runtime value of type RUNTIME_VALUE_SYMBOL,
holding the symbol pointer itself. This caused incorrect behavior, as symbols
should normally be *evaluated* by resolving their bound value in the runtime
environment.

Now, the interpreter resolves symbol nodes using runtime_env_get(), which
returns the runtime_env_value currently bound to that symbol. The retrieved
value is now retained (reference-counted) before being returned through *out*,
ensuring safe lifetime management even if the originating environment frame
is later unwound. This change is part of a broader migration toward a
retain/release ownership model for runtime_env_value instances, replacing the
previous copy-based semantics.

This change also introduces const-correctness across the symbol evaluation
path: runtime_env_get() and related APIs now return `const runtime_env_value *`
to make explicit that the returned value must not be modified directly.
This adjustment clarifies ownership and immutability semantics, enforcing the
retain/release model for runtime_env_value instances and replacing earlier
copy-based behavior.

If the symbol is unbound, the evaluation fails with
INTERPRETER_STATUS_UNBOUND_SYMBOL.

Certain nodes intentionally bypass symbol resolution:
  - AST_TYPE_QUOTE: symbols are treated as literal values, not looked up.
  - AST_TYPE_BINDING (and similar LHS constructs): symbols denote identifiers
    being defined, not values to be resolved.

These adjustments make symbol evaluation consistent with expected interpreter
semantics, clarify ownership rules for runtime values, and prevent invalid
memory accesses during environment unwinds.

+++ src/runtime_env/include/internal/runtime_env_internal.h
@@ -16,12 +16,10 @@ void runtime_env_value_destroy_adapter(void *value);
 
 struct runtime_env {
     hashtable *bindings; // key: const symbol* (borrowed), value: runtime_env_value (owned)
-    int refcount;
+    size_t refcount;
     bool is_root;
     struct runtime_env *parent;
 };
-
-// make the following fcts static?
 
 runtime_env *runtime_env_make_toplevel(void);
 
@@ -29,5 +27,7 @@ runtime_env_value *runtime_env_value_clone(const runtime_env_value *value);
 void runtime_env_destroy(runtime_env *e);
 void runtime_env_retain(runtime_env *e);
 void runtime_env_release(runtime_env *e);
+
+//void runtime_env_value_destroy(runtime_env_value *value);
 
 #endif //LEXLEO_RUNTIME_ENV_INTERNAL_H
